MAPLE-13 멀티모듈로 구성
