+++ api/src/main/java/com/maple/api/MapleBossApiApplication.java
@@ -1,0 +1,12 @@
+package com.maple.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MapleBossApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MapleBossApiApplication.class);
+    }
+}
