+++ common/src/main/java/com/maple/common/boss/domain/BossClass.java
@@ -1,0 +1,5 @@
+package com.maple.common.boss.domain;
+
+public enum BossClass {
+    EASY
+}
