+++ common/src/main/java/com/maple/common/boss/domain/BossRepository.java
@@ -1,0 +1,6 @@
+package com.maple.common.boss.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BossRepository extends JpaRepository<Boss, Long> {
+}
