+++ common/src/main/java/com/maple/common/boss/service/BossService.java
@@ -1,0 +1,7 @@
+package com.maple.common.boss.service;
+
+import com.maple.common.boss.domain.Boss;
+
+public interface BossService {
+    Boss create(Boss boss);
+}
