+++ common/src/main/java/com/maple/common/support/BaseEntity.java
@@ -1,0 +1,18 @@
+package com.maple.common.support;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+
+@Getter
+@EqualsAndHashCode
+@MappedSuperclass
+public abstract class BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+}
