+++ common/src/main/java/com/maple/common/support/Constants.java
@@ -1,0 +1,5 @@
+package com.maple.common.support;
+
+public final class Constants {
+    public static final String COMMON_BASE_PACKAGE = "com.maple.common";
+}
