+++ common/src/test/java/com/maple/common/MapleBossCoreApplication.java
@@ -1,0 +1,7 @@
+package com.maple.common;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+class MapleBossCoreApplication {
+}
