+++ common/src/test/java/com/maple/common/boss/service/DomainBossServiceTest.java
@@ -1,0 +1,30 @@
+package com.maple.common.boss.service;
+
+import com.maple.common.boss.domain.BossRepository;
+import com.maple.common.fixture.BossFixture;
+import com.maple.common.support.BaseServiceTest;
+import lombok.val;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class DomainBossServiceTest extends BaseServiceTest {
+
+    @Autowired
+    private BossRepository bossRepository;
+
+    @Autowired
+    private BossService bossService;
+
+    @Test
+    void 보스_생성_성공() {
+        var boss = BossFixture.createBoss();
+
+        boss = bossService.create(boss);
+
+        val foundBoss = bossRepository.findById(boss.getId()).orElseThrow();
+
+        assertThat(foundBoss).isEqualTo(boss);
+    }
+}
