+++ common/src/test/java/com/maple/common/support/BaseServiceTest.java
@@ -1,0 +1,11 @@
+package com.maple.common.support;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
+
+import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.NONE;
+
+@Transactional
+@SpringBootTest(webEnvironment = NONE)
+public abstract class BaseServiceTest {
+}
