+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'maple-boss-server'
+
+include 'common'
+include 'api'
