--- src/main/java/com/maple/MapleBossServerApplication.java
@@ -1,13 +1,0 @@
-package com.maple;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MapleBossServerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MapleBossServerApplication.class, args);
-    }
-
-}
