--- src/main/java/com/maple/domain/boss/domain/BossClass.java
@@ -1,5 +1,0 @@
-package com.maple.domain.boss.domain;
-
-public enum BossClass {
-    EASY
-}
