--- src/main/java/com/maple/domain/boss/domain/BossRepository.java
@@ -1,6 +1,0 @@
-package com.maple.domain.boss.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BossRepository extends JpaRepository<Boss, Long> {
-}
