--- src/main/java/com/maple/domain/boss/service/BossService.java
@@ -1,7 +1,0 @@
-package com.maple.domain.boss.service;
-
-import com.maple.domain.boss.domain.Boss;
-
-public interface BossService {
-    Boss create(Boss boss);
-}
