--- src/main/java/com/maple/support/BaseEntity.java
@@ -1,18 +1,0 @@
-package com.maple.support;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-
-@Getter
-@EqualsAndHashCode
-@MappedSuperclass
-public abstract class BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-}
