--- src/test/java/com/maple/MapleBossServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.maple;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MapleBossServerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
