--- src/test/java/com/maple/support/BaseServiceTest.java
@@ -1,11 +1,0 @@
-package com.maple.support;
-
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.NONE;
-
-@Transactional
-@SpringBootTest(webEnvironment = NONE)
-public abstract class BaseServiceTest {
-}
